requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
databases==0.5.5
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.7.1
fastapi==0.75.2
fastapi-mail==1.0.8
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
mutagen==1.45.1
ormar==0.11.0
packaging==21.3
passlib==1.7.4
Pillow==9.1.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.8
python-jose==3.3.0
python-multipart==0.0.5
redis==4.1.4
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing_extensions==4.2.0
uvicorn==0.17.6
wrapt==1.14.0
